requirements.txt
google-genai
flask
Flask-RESTful
dotenv
simplegmail
google-auth-oauthlib
google-api-python-client
requests
